Status Checking fixed
