--- create_db.py
@@ -1,4 +1,0 @@
-from nova_optuna import create_study_with_storage
-
-if __name__ == "__main__":
-    create_study_with_storage()
